Reorganize code into src/ and tests/ folders
- Move lino-env.mjs to src/ folder for better organization
- Move lino-env.test.mjs to tests/ folder
- Update package.json main and exports paths to src/lino-env.mjs
- Update jest.config.js to use <rootDir> paths
- Update examples/basic-usage.mjs import path
- Update GitHub Actions workflow Deno import path
- All tests passing (22/22)
- All lint and format checks passing
